test ICU

+++ HAL/stepper_motor/stepper_interface.h
@@ -5,8 +5,8 @@
 #define HALF_STEP   8
 
 void Stepper_voidInit(void);
-void stepper_Rotate_CW(u8 Copy_StepMode, u16 degree );
-void stepper_Rotate_CCW(u8 Copy_StepMode, u16 degree );
+void stepper_Rotate_CW(u8 Copy_StepMode, f32 degree );
+void stepper_Rotate_CCW(u8 Copy_StepMode, f32 degree );
 
 
 #endif
